Plantas muestran timer 00:00 y pueden ser cosechadas manualmente una a una. Se liberan correctamente los slots después, y se agregan los frutos al inventario.

+++ Assets/Scripts/BarraEnergiaSistema.cs
@@ -159,7 +159,7 @@ public class BarraEnergiaSistema : MonoBehaviour
     /// <summary>
     /// RQNF53.2: Calcula el tiempo de crecimiento modificado según energía
     /// </summary>
-    public int GetModifiedGrowthTime(int baseTimeSeconds)
+    public float GetModifiedGrowthTime(float baseTimeSeconds)
     {
         int penalty = GetTimePenalty();
         return baseTimeSeconds + penalty;
